requirements.txt
groq>=0.9.0
chromadb>=0.5.3
sentence-transformers>=3.0.1
transformers>=4.41.2
torch
detoxify>=0.5.2
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.1
python-dotenv>=1.0.1
langgraph>=0.1.14
scikit-learn>=1.5.0
requests>=2.32.3
mysql-connector-python>=9.0.0
